Reorganize wookie middleware/context methods
