+++ pkg/wookie/wookie.go
@@ -16,34 +16,15 @@ package wookie
 
 import (
 	"context"
-	"net/http"
 )
 
-const HeaderName = "Warrant-Token"
 const Latest = "latest"
 
-type wookieCtxKey struct{}
-
-func WookieMiddleware(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		headerVal := r.Header.Get(HeaderName)
-		if headerVal == Latest {
-			latestCtx := WithLatest(r.Context())
-			next.ServeHTTP(w, r.WithContext(latestCtx))
-			return
-		}
-		next.ServeHTTP(w, r)
-	})
-}
-
-// Return a context with wookie set to 'latest'.
-func WithLatest(parent context.Context) context.Context {
-	return context.WithValue(parent, wookieCtxKey{}, Latest)
-}
+type WookieCtxKey struct{}
 
 // Returns true if ctx contains wookie set to 'latest', false otherwise.
 func ContainsLatest(ctx context.Context) bool {
-	if val, ok := ctx.Value(wookieCtxKey{}).(string); ok {
+	if val, ok := ctx.Value(WookieCtxKey{}).(string); ok {
 		if val == Latest {
 			return true
 		}
